FHEW DM done
